patients/requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
appdirs==1.4.3
asgiref==3.2.10
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.0
behave==1.2.6
boto==2.49.0
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
distlib==0.3.0
dj-database-url==0.5.0
Django==2.1.15
django-extensions==2.2.9
edx-dl==0.1.13
filelock==3.0.12
Flask==1.1.2
gast==0.2.2
get==2019.4.13
gevent==1.5.0
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
langdetect==1.0.8
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
multidict==4.7.5
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
parse==1.14.0
parse-type==0.5.2
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
python-decouple==3.3
pytz==2020.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
toml==0.10.1
Unipath==1.1
urllib3==1.25.8
virtualenv==20.0.17
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
yarl==1.4.2
youtube-dl==2020.3.24
zipp==3.1.0
